refactor(tunnel): split relay lifecycle behind platform seams
